quantum_computer/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dill==0.4.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.2
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.0
fastar==0.9.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
ibm-cloud-sdk-core==3.24.4
ibm-platform-services==0.74.0
ibm-quantum-schemas==0.5.20260320
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.4.3
orjson==3.11.7
packaging==26.0
psutil==7.2.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.12.1
pymongo==4.16.0
pyspnego==0.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
qiskit==2.3.1
qiskit-aer==0.17.2
qiskit-ibm-runtime==0.46.1
requests==2.32.5
requests_ntlm==1.3.0
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
rustworkx==0.17.1
samplomatic==0.17.1
scipy==1.17.1
sentry-sdk==2.56.0
shellingham==1.5.4
six==1.17.0
sspilib==0.5.0
starlette==1.0.0
stevedore==5.7.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0
